Removed Onboarding as we will use parts of the settings screen. Fixed SettingsFragments conversions.

+++ app/src/main/java/com/iate/android/data/database/entity/UserSettings.kt
@@ -9,8 +9,8 @@ data class UserSettings(
     val metric: Boolean,
     val age: Int,
     val gender: String,
-    val height: Int,
-    val weight: Int,
+    val height: Double, // Changed from Int to Double
+    val weight: Double, // Changed from Int to Double
     val activityLevel: String,
     val bmr: Int,
     val tdee: Int,
